TASK: Move `TranslationAddress` into `model` module
