+++ packages/neos-ui-i18n/src/registry/index.ts
@@ -14,5 +14,4 @@ export type {Parameters} from './Parameters';
 
 export {substitutePlaceholders} from './substitutePlaceholders';
 
-export {TranslationAddress} from './TranslationAddress';
 export {TranslationRepository} from './TranslationRepository';
